refactor and update message dialog
